Compatibility fixes for Lua-5.2.x (compiles and runs with 5.2.3)
You should now be able to build FXite with your system's Lua 5.2 shared
library. Lua 5.1 is still the default, and Lua 5.1.5 is still what's
included with the FXiTe sources. There is currently no fully automatic
way to tell configure to use 5.2, so you will need to find the name of
your 5.2 pkg-config script and pass its basename to configure with the
--with-lua-pkg= parameter, along with the --with-system-lua flag.
I was able to build on ArchLinux using this command:

  ./configure --with-system-lua --with-lua-pkg=lua

Of course, this being brand new code it needs all the testing it can
get, and it is even possible that I introduced some bugs in the old
Lua 5.1 bindings as well.